policy file was delete right after being created
This caused similar issues as report in in #33 albeit, it can't be the
root cause for that particular report as the change that caused this was
done after the bug was reported.

Anyway, the bug here was that as there was no existing policy file,
javaagent didnt have permissions to connect RSL server, thus causing
agent timeout.